+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 ## [Unreleased]
 
+- Optimize aarch64 128-bit load when the `lse` target feature is enabled at compile-time.
+
 ## [0.2.11] - 2022-07-31
 
 - Optimize inline assemblies on x86, x86_64, ARM, AArch64, and RISC-V.